+++ example/lib/containers/form_container.dart
@@ -79,6 +79,7 @@ class TFormContainerExample extends StatelessWidget {
         TFormContainer(
           isLoading: false,
           showShadow: true,
+          fullWidth: true,
           child: Column(
             mainAxisSize: MainAxisSize.min,
             children: [